Fix gmapping depthimage_to_laserscan nodelet problem
The gmapping_demo.launch turtlebot_navigation calls turtlebot_bringup/launch/3dsensor.launch in which there was a nodelet for converting depth-image to laser-scan (and thus using a 3D rgbd camera as a lidar). This nodelet has been changed to a normal package node, because it did not work as nodelet.
Also, there was a remapping wrong, because our D435 camera generates /image_rect_raw instead of the usual /image_raw topic.